requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
httpx
chromadb
python-dotenv
tiktoken
numpy
python-multipart